+++ src/treasure_map/agent/__init__.py
@@ -1,2 +1,0 @@
-# Copyright (C) 2025-2026 JoeyZzZzZz
-# SPDX-License-Identifier: AGPL-3.0-only